requirements.txt
numpy~=1.19.5
matplotlib~=3.3.3
openpyxl~=3.0.6
deap~=1.3.1
sklearn~=0.0
scikit-learn~=0.23.2
Pillow~=8.1.0
pandas~=1.2.0
bayesian-optimization~=1.2.0
scipy~=1.5.4
scikit-optimize~=0.8.1
gym~=0.20.0